--- main.cpp
@@ -1,9 +1,0 @@
-#include <string>
-
-int main() {
-
-    std::cout << "Hi!" << std::endl;
-    dsda
-
-    return 0;
-}
