+++ src/lex.cpp
@@ -1,4 +1,4 @@
-#include "lib/lex.h"
+#include "lib/Lexer.h"
 
 int main() {
 
@@ -13,14 +13,6 @@ int main() {
         input += line;
     }
     input = input.substr(0, input.size()-1);
-
-    // std::cout << "hello world" << std::endl;
-    // // Token::printLexer("(+(-2 4.444 )\n32(* 5 13.45)(");
-    // Token::printLexer("(+(-2 4.444 )");
-    // // Token::printLexer("(/49 50.\n12 ($ 7..8 .3))");
-    // return 0;
-    
-
 
     std::vector<Token> TokenVector = Token::GenTokenVector(input);
     // Check ERROR
