+++ src/lib/Lexer.cpp
@@ -1,49 +1,4 @@
-#pragma once
-
-#include <string>
-#include <vector>
-#include <iostream>
-#include <cmath>
-
-enum class TokenType {
-    none,
-    number,
-    plus,
-    minus,
-    multiply,
-    divide,
-    leftParenthesis,
-    rightParenthesis,
-    end,
-    error
-};
-
-class Token {
-
-public:
-    Token(TokenType itype, std::string icontent, int iline, int iindex, double ivalue = -1);
-    ~Token() = default;
-
-    operator double() const {
-        return value;
-    }
-
-public:
-    static std::vector<Token> GenTokenVector(const std::string input);  // returns a vector of tokens created from a string.
-    static void printLexer(std::vector<Token> TokenVector);
-    static void printLexer(const std::string input);
-     
-public:
-    TokenType type;
-    std::string content;    // value as a string
-    int line = -1;
-    int index = -1;
-    double value = -1;       // a number has its own value. otherwise -1
-
-};
-
-
-
+#include "Lexer.h"
 
 Token::Token(TokenType itype, std::string icontent, int iline, int iindex, double ivalue)
     : type(itype), content(icontent), line(iline), index(iindex), value(ivalue)
@@ -62,23 +17,24 @@ std::vector<Token> Token::GenTokenVector(const std::string input) {
 
     // for number
     int numberLength = 0;
-    double currentValue = 0;         
+    double currentValue = 0;
     std::string currentStringValue = "";
-    double afterPoint = 0;   
+    double afterPoint = 0;
     bool recordingNumber = false;
 
     for (int i = 0; i < len; i++)
     {
         // std::cout << (int)input.at(i) << std::endl;
-        if (input.at(i) == '0' || input.at(i) == '1' || input.at(i) == '2' || input.at(i) == '3' || input.at(i) == '4' || input.at(i) == '5' || input.at(i) == '6' || input.at(i) == '7' || input.at(i) == '8' || input.at(i) == '9') 
+        if (input.at(i) == '0' || input.at(i) == '1' || input.at(i) == '2' || input.at(i) == '3' || input.at(i) == '4' || input.at(i) == '5' || input.at(i) == '6' || input.at(i) == '7' || input.at(i) == '8' || input.at(i) == '9')
         {
             recordingNumber = true;
             if (afterPoint) {
                 currentValue += double(input.at(i) - '0') / std::pow(10.0f, afterPoint);
                 afterPoint += 1;
-            } else {
+            }
+            else {
                 currentValue *= 10;
-                currentValue += double(input.at(i) - '0');   
+                currentValue += double(input.at(i) - '0');
             }
             std::string s(1, input.at(i));
             currentStringValue += s;
@@ -161,7 +117,7 @@ std::vector<Token> Token::GenTokenVector(const std::string input) {
                 res.emplace_back(TokenType::rightParenthesis, ")", line, index, -1);
             }
             // Error
-            else    
+            else
             {
                 // if there's error, the last token would be ERROR instead of END
                 res.emplace_back(TokenType::error, "ERROR", line, index, -1);
@@ -193,7 +149,7 @@ std::vector<Token> Token::GenTokenVector(const std::string input) {
 }
 
 
-void Token::printLexer(std::vector<Token> TokenVector) 
+void Token::printLexer(std::vector<Token> TokenVector)
 {
 
 
@@ -213,7 +169,7 @@ void Token::printLexer(std::vector<Token> TokenVector)
         // line
         content = std::to_string(t.line);
         spaceNumber = 4 - content.length();
-        for (int i=0; i<spaceNumber; i++)
+        for (int i = 0; i < spaceNumber; i++)
         {
             std::cout << " ";
         }
@@ -222,7 +178,7 @@ void Token::printLexer(std::vector<Token> TokenVector)
         // index
         content = std::to_string(t.index);
         spaceNumber = 5 - content.length();
-        for (int i=0; i<spaceNumber; i++)
+        for (int i = 0; i < spaceNumber; i++)
         {
             std::cout << " ";
         }
@@ -239,7 +195,7 @@ void Token::printLexer(std::vector<Token> TokenVector)
 
 
 
-void Token::printLexer(const std::string input) 
+void Token::printLexer(const std::string input)
 {
     std::vector<Token> TokenVector = GenTokenVector(input);
     Token::printLexer(TokenVector);
